Build command working. Added several other supporting moduiles, extensive refactoring, templates, etc...

+++ src/context/mod.rs
@@ -6,7 +6,7 @@ use crate::config::EchoConfig;
 #[derive(Serialize, Deserialize, Default)]
 pub struct EchoContext {
     pub config: EchoConfig,
-    pub num_posts: u64,
+    pub num_posts: isize,
     pub posts: Vec<EchoPost>
 }
